server/requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
opencv-python==4.8.1.78
numpy==1.24.3
ultralytics==8.0.200
Pillow==10.1.0
python-engineio==4.8.0
eventlet==0.33.3